Separate out config handling to separate class
